Prevent #merge_target_lists from clearing id column on CPK records
Previously, when merging persisted and in-memory records, we were using
`record[name]` to get the value of the attribute. For records with a
composite primary key, the full CPK would be returned, which was
incompatible with `#_write_attribute`: we'd attempt to assign an Array
to the `id` column, which would result in the column being set to nil.

When dealing with a record with a composite primary key, we need to
write each part of the primary key individually via `#_write_attribute`.