Add Rack::Lint to ActionableExceptions tests
This adds additional test coverage to ActionableExceptions to validate
that its behavior conforms to the Rack SPEC.

The changes neccesary were to ensure that Response headers are downcased
when using Rack 3. For Content-Type and Content-Length, this is trivial
because Rack provides constants who's casing is dependent on the version
(Rack 2 is mixed, and Rack 3 is downcased). Since Rack does not include
a LOCATION constant, the Response::LOCATION constant was updated to
have a downcased value when using Rack 3.

Additionally, there was some missing coverage for invalid redirect URLs
which was addressed as well.