Revert "Merge pull request #49242 from eileencodes/fix-inferred-query-constraints"
This reverts commit 213e8ff95e71cd35b9ae51fcbc26e97ed8eda658, reversing
changes made to b5d63b9b16c6f251eaa08f61086f21ff3ef3292f.

It broke a lot of test.

It also changed how lazy loading of model code works.

With this change a simple `belongs_to :owner` is now loading `owner.rb`
when that line is executed, where before it wasn't doing that.
This can have serious impact on boot time in development.

We should make this as lazy as possible.