Get rid of ActiveRecord::Core#===
This was added back in 0.9.5 because at the time single associations (e.g. has_one)
didn't directly return the record but a proxy

https://github.com/rails/rails/commit/97849debf33

But toaward 3.1, we got rid of that proxy: https://github.com/rails/rails/commit/1644663ba7f678d178deab2bf1629dc05626f85b
and the associated test was deleted.

However the `===` method stayed.

This has been around for so long that removing it may
break some rare applications, but I doubt it would be hard to
fix, and this isn't a documented behavior.

Co-Authored-By: Nate Matykiewicz <natematykiewicz@gmail.com>