Fix `ActionController::Parameters#deep_merge` RDoc [ci skip]
Follow-up to [#45369][]

First, add the final `:` to the `#deep_merge?` method's `:nodoc:`
declaration.

Next, move the `:method:` documentation out of the methods and into the
section of the class that defines the rest of the dynamic method
documentation.

Finally, move the `:call-seq:` methods below the directive, like the
rest of the methods.

[#45369]: https://github.com/rails/rails/pull/45369