Add materialized? to NullTransaction
The NullTransaction implements the null object pattern for the
Transaction class, but it was missing the `materialized?` method.

This caused issues in cases where we'd want to check if the current
transaction is materialized, where we'd need to check if the connection
is open first:

```ruby
current_transaction = ActiveRecord::Base.connection.current_transaction
current_transaction.open? && current_transaction.materialized?
```

With this change we can skip the extra check on `open?`

```ruby
current_transaction = ActiveRecord::Base.connection.current_transaction
current_transaction.materialized?
```