Optimize `ActiveRecord::QueryMethods.excluding`
Prior to this commit, passing an `ActiveRecord::QueryMethods` or
`ActiveRecord::Relation` **was** supported, but it was effective only
because its records would be coerced into an `Array` through the
splatting, `flatten!`-ing, and `compact!`-ing.

This commit adds explicit support, and optimizes the underlying query by
omitting all fields that aren't the table's primary key.

Co-authored-by: Jonathan Hefner <jonathan@hefner.pro>