Fix casting PG money with comma as radix point
The `/^-?\D*+[\d.]+,\d{2}$/` regexp in the guard does not guarantee a
match for the subsequent `gsub!`.  For example, when casting `"3,50"`.
Thus `gsub!` could return `nil`, and the chained `sub!` call would raise
`NoMethodError`.

This commit splits up the call chain.  It also replaces `gsub!` with
`delete!` and `sub!` with `tr!` since regexps are unnecessary for these
cases.

Co-authored-by: Jonathan Hefner <jonathan@hefner.pro>