Recover from failed connections in PostgreSQL
Fixes https://github.com/rails/rails/issues/49781

Prior to this commit, the PostgreSQL adapter could end up holding onto a
broken connection longer than expected, potentially **into the next
request after the executor checkin/checkout**.

The problem is that the PostgreSQL adapter was using a [two separate
calls] to `with_raw_connection` for preparing vs. executing a prepared
statement. That's fine when the statement is first prepared, but once it
is in the statement pool the first call to `with_raw_connection` [does
not actually use the connection]. But that doesn't stop it from marking
the connection as [`@verified`][verified], preventing the second
`with_raw_connection` from verifying and getting a chance to reconnect.

[two separate calls]: https://github.com/rails/rails/blob/e33dbe1b3d423143685154b97ae0992f4ab821b8/activerecord/lib/active_record/connection_adapters/postgresql_adapter.rb#L910-L917
[does not actually use the connection]: https://github.com/rails/rails/blob/490b9fc22faefca30565a970a8727c915de589ff/activerecord/lib/active_record/connection_adapters/postgresql_adapter.rb#L964-L966
[verified]: https://github.com/rails/rails/blob/e33dbe1b3d423143685154b97ae0992f4ab821b8/activerecord/lib/active_record/connection_adapters/abstract_adapter.rb#L1032

```rb
Post.first
=> prepare_statement verifies the connection (if needed) and populates the statement pool
=> exec_prepared succeeds

=> Oops!, we lost the connection

Post.first
=> prepare_statement returns the statement from the statement pool, connection still marked as verified
=> exec_prepared fails, marks the connection as unverified

Post.first
=> prepare_statement returns the statement from the statement pool, marks the connection as verified!!!
=> exec_prepared fails, marks the connection as unverified

=> request finishes, connection gets marked as unverified for the next request

Post.first
=> prepare_statement returns the statement from the statement pool, marks the connection as verified!!!
=> exec_prepared fails, marks the connection as unverified
```

The only way to recover is to execute a statement that hasn't been
prepared yet, one that isn't prepared because it has no binds, or one
that has `allow_retry: true` (which I think is limited to a few internal
queries at the moment—we don't pass the value through to
`with_raw_connection` for prepared statements anyway).

This commit fixes the problem by preparing and executing the statement
within a single `with_raw_connection` block. This is what the other
adapters with prepared statements already do.

It makes more sense this way—even if we managed to reconnect before
executing the statement in the old code, we'd be referencing a prepared
statement from a different connection, so it'd fail anyway. Having the
prepare and execute in the same `with_raw_connection` means if we do
need to reconnect, we'll start from the beginning and prepare the
statement again on the fresh connection.

Note that this change does get rid of connection retries on preparing
statements, but I think that's OK. It matches the other adapters,
matches the Rails 7.0 behavior, and it's still better than ending up
with a broken connection.