Prebuild list of counter cache associations
Currently when we _create_record or destroy_row
we have to iterate over the list of _reflections
and find the ones that are belongs_to type and have a counter_cache_column.
This is inefficient as we don't expect the result of this selection to change during runtime.
To make it prettier and more efficient we will
prebuild a list of associations that have a counter cache column.