Add caveat about normalizes / Marshal interaction [ci-skip]
When `ActiveRecord.marshalling_format_version` is set to `6.1`,
`Marshal` will try to serialize attribute types along with the model,
causing a `TypeError` if the model uses `ActiveRecord::Base.normalizes`
with a normalization `Proc`.

This commit adds a caveat to the `normalizes` API documentation to warn
users that they should set `marshalling_format_version` to `7.1` if they
are using `normalizes` and marshalling the targeted model.

Fixes #49871.