Clarify transaction callbacks section [ci-skip]
Rename "Dealing With Consistency" heading to "`after_commit` and
`after_rollback`. To be more inline with other headings.

The "Context Matters" heading has been renamed to "Aliases for
`after_commit`", as this section is about aliases and `Context Matters` is
a bit ambiguous.

The warnings for `after_commit` and `after_rollback` better fit with the
section about these callback methods.