Add connection pool to connection timeout errors
Along the same lines as https://github.com/rails/rails/pull/48295,
in multi-database applications it's helpful to have access to the pool
when debugging these timeouts. For example, we might send the connection
class and role along to an exception tracking system.

ConnectionTimeout was already set up to take a connection_pool, so this
commit passes it along as needed where we raise the error.