Implement `CurrentAttributes#set` in terms of `Object#with`
`CurrentAttributes` supports block-scoped overrides for its attributes
through the `#set` method. The introduction of [CurrentAttributes#set][]
predates the introduction of [Object#with][] by 6 years.

This commit changes the implementation of `#set` to delegate to `#with`.
Through that delegation, the private `#assign_attributes` and
`#compute_attributes` methods are no longer necessary.

[CurrentAttributes#set]: https://github.com/rails/rails/blame/2d6b02bad6b02bf1d26c26461406bda710181244/activesupport/lib/active_support/current_attributes.rb#L210
[Object#with]: https://github.com/rails/rails/blame/2d6b02bad6b02bf1d26c26461406bda710181244/activesupport/lib/active_support/core_ext/object/with.rb#L26