Unify the logic to determine the default schema cache path for a database configuration
When the application has more than one database configuration, only
the primary was being loaded by default on boot time. All the other
connection pools were loading the schema cache lazily.

This logic can be found in:

https://github.com/rails/rails/blob/351a8d9bc99e69cc8d1ced86dc1d523e2dc51985/activerecord/lib/active_record/connection_adapters/pool_config.rb#L13
https://github.com/rails/rails/blob/351a8d9bc99e69cc8d1ced86dc1d523e2dc51985/activerecord/lib/active_record/railtie.rb#L149-L178

The lazy code path wasn't using the same logic to determine the name
of the default schema cache file that the Railties uses, so it
was loading the schema cache of the primary config to all the other
configs. If no table name coincided nothing bad would happen, but if
table names coincided, the schema would be completly wrong in all
non-primary connections.