Don't query adapter in ensure blocks when connection is closed
Trilogy closes its connection when it encounters a timeout. Active
Record doesn't always account for this, and can query the adapter in
ensure blocks of certain adapter methods. Instead, we need to preface
these queries with active checks to make sure we don't swallow the
timeout and raise an adapter error instead.

Co-authored-by: Adrianna Chang <adrianna.chang@shopify.com>