Extract check for VALUES/row_alias syntax
The change to upserts (https://github.com/rails/rails/pull/51274) is causing
an issue with Vitess, since it doesn't support the `row_alias` syntax added
in MySQL 8.0.19.

There is an ongoing work (https://github.com/vitessio/vitess/pull/15510)
to add that support and it is likely to be included into Vitess v20,
but in the meantime it would be nice to have an ability to control that
behavior in the existing apps.