Revert "Eliminate lease_connection call in TypeCaster::Connection"
This reverts commit 6dd1929b04de42cdff1264bb5618c7f6abe77a9c.

This introduced a change in behaviour since type_for_attribute is aware of custom types but lookup_cast_type does not.

Additionally, since we no longer to use the table and column info to get attributes, this introduced an issue where attribute types were not be correctly found for some queries, where we were joining a table that has a different name than the name of the reflection for that association.